Add real EGL handle tracking mechanism
Add a tsearch()-based binary tree of valid EGL
handles and a mutex to protect use of it. Man
pages claim the tsearch() family of functions are
thread-safe, but the way it is used here requires
critical sections larger than a single glibc call
unfortunately, so a separate mutex is required.