Implement eGbmIsValidNativeDisplayExport()
Checks the specified nativeDpy to see if it is a
GBM device if possible. Uses mincore to inspect
the object pointed to if possible, and accepts
an environment variable override (EGL_PLATFORM).
If unable to inspect the pointer and no override
is specified, it is assumed the nativeDpy is not
a GBM device.