Add EGL_DEFAULT_DISPLAY handling
If the default device is reqeusted, try to open
the first enumerated DRM device in the system,
create a GBM device from it, and use that as the
default native display.

Note only the first device is attempted. If this
fails, it is better to return failure back up to
GLVND rather than walk all the devices hoping one
succeeds. GLVND can instead potentially try the
request on other EGL implementations, which may
be able to support the first enumerated DRM
device. This will hopefully ensure stable default
device selection regardless of EGL driver vendor
library load order within GLVND.