Support creating EGLSurfaces from gbm_surfaces
Plug in the necessary functions for creating EGL
window surfaces from struct gbm_surface objects.
The surfaces are backed by an EGLStream producer
surface with an EGLImage consumer. The stream
will produce images, from which the surface
code will construct and import GBM bo's, which
in turn will be returned to the application when
it calls gbm_surface_lock_front_buffer().

There is currently no cleanup code, so the BOs and
EGL objects are leaked when the surface and
EGLSurface are destroyed.