Add EGL_WINDOW_BIT EGLConfigs for gbm_surface
Now that there is support for creating a window
surface from a gbm_surface object + an EGLStream,
expose the EGL_WINDOW_BIT surface type for each
EGLConfig that supports creating an EGLStream
producer surface. The code here was largely based
on similar code in the EGL Wayland external
platform library:

  https://github.com/NVIDIA/egl-wayland

This is not yet entirely accurate. Each EGLConfig
should also likely be inspected to ensure it is
compatible with some GBM format supported by the
GBM device first, and the native visual attribute
should also likely be taken into account (In the
form of a GBM format/drm_fourcc code).

However, this should be enough to start testing
GBM-based EGL applications that render to an
EGLSurface, such as kmscube.