Use an EGLSync object to wait for stream frames
When acquiring an eglimage from an EGLStream, an
EGLSync object can be provided to wait for
any pending work work the producer has created
utilizing the image. Create a reusable EGLSync
associated with each gbm_surface EGLSurface, pass
it in when acquiring frames, and immediately wait
for it to idle the surface before handing it off
to the application. This will ensure the surface
is idle until we can add true sync FD/explicit
sync support on desktop GPUs.