EGL/GBM: Partially fix eGbmSurfaceHasFreeBuffers.
Add a counter for the number of free color buffers to GbmSurface. That
counter is initialized to the stream's FIFO length, decremented when we
acquire an EGLImage from the stream, and incremented when we release an
EGLImage.

Changed eGbmSurfaceHasFreeBuffers to return non-zero if the
GbmSurface::numFreeImages counter is non-zero.

Renamed GbmSurface::freeImages to acquiredImages, since that's the list
of EGLImages that we've acquired, and thus, the images that are *not*
free.

Note that the behavior of eglSwapBuffers is still wrong, and things will
break if an application tries to call eglSwapBuffers without calling
gbm_surface_lock_front_buffer afterward. Fixing that will require more
extensive changes.