egl-gbm: Return early for non-NVIDIA devices.
In eGbmGetPlatformDisplayExport, check the backend driver name, and
return early if that's not "nvidia".

If the gbm_device doesn't correspond to an NVIDIA device, then this lets
it return early, before calling eglQueryDevicesEXT. That's important on
multi-GPU systems, where calling eglQueryDevicesEXT might have to power
on an NVIDIA GPU, which wastes power and can be very slow.