Fix a segfault if drmGetDevices2 fails.
In OpenDefaultDrmDevice, return early if drmGetDevices2 returns less
than or equal to zero, not just zero.

If drmGetDevices2 fails (e.g., if /dev/dri doesn't exist yet), then
it'll return a negative errno value, in which case the current code
would continue on and try to dereference the still-uninitialized
drmDevicePtr.